Splitting out the managers into different files.
